Added additional class that handles spawning tasks and employed singleton design on pedometer service.
